Update test network (docs, package, compose file)
I have added the current version of the packaged chaincode (tpcc.tgz) to
the directory where the peers will actually be able to install chaincode
from in the Fablo network (this can be done using the copy_to_fablo
script).

This commit also updates the main test network README and sets the
correct package ID wherever it appears.