Get gradle shadowJar working with OpenJML build
Things are expected to work in Docker, when the CC image is built using
the Dockerfile in its directory.  You might (and probably will)
experience problems when trying to build this locally.  This is because
of potential the Gradle version and JDK version inconsistencies.

buildSrc/ now contains a bit of logic to download and set up compilation
with OpenJML, essentially creating a copy of the JDK Gradle auto-locates
and overwriting the javac and java binaries there with ones that use
OpenJML's JAR instead.

The resulting Docker image should be self-contained CCAAS chaincode,
ready to be used.